Moved common/shared C++ code to reprostim::CaptureLib library #61
